+++ CHANGELOG.md
@@ -13,6 +13,10 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - Server: Allow CLI to update app block themes if it has the `apps:write` scope.
 - Server: Fix issue with invalid role validation for pages with sub pages.
 
+### Changed
+
+- SDK: Remove `block` from bootstrap parameters, `parameters` is now passed in directly.
+
 ## [0.11.6] - 2020-03-05
 
 ### Added